requirements-dev.txt
-r requirements-common.txt
pytest
fastapi
uvicorn[standard]
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas